Pushing Finished Sim Exec and Updated Warehouse

+++ SWAS/Warehouse.h
@@ -4,11 +4,11 @@
 class Warehouse
 {
 public:
-	Warehouse(int length, int width);
-	int GetLength();
-	void SetLength(int hangarLength);
-	void SetWidth(int hangarWidth);
-	int GetWidth();
+	Warehouse(double length, double width);
+	double GetLength();
+	void SetLength(double hangarLength);
+	void SetWidth(double hangarWidth);
+	double GetWidth();
 private:
 	double _hangarLength;
 	double _hangarWidth;
